fix graceful
"apachctl graceful" was failing: we'd lock during restart.

This extension was being unloaded, which in turn unloaded libvips (since
only we referenced it), but glib was not being unloaded, since other
things used by apache were still using it.

As a result, on reload of this extension, libvips would try to init, and
discover that parts of its previous init were still there, such as the
VipsObject type.

Fix: lock libvips in memory to prevent unload.

See https://github.com/jcupitt/php-vips/issues/26